Added message write endpoint

+++ discordcs.core/src/Interfaces/Channel/Message/IAttachment.cs
@@ -6,7 +6,7 @@ namespace Discordcs.Core.Interfaces
 		public string Filename { get; set; }
 		public string Description { get; set; }
 		public string ContentType { get; set; }
-		public uint Size { get; set; }
+		public long Size { get; set; }
 		public string Url { get; set; }
 		public string ProxyUrl { get; set; }
 		public uint? Height { get; set; }
